Update best_answer_id in questions table when answer deleted in database layer

+++ database/migrations/2019_06_17_051940_create_questions_table.php
@@ -24,7 +24,7 @@ class CreateQuestionsTable extends Migration
             $table->integer('views')->unsigned()->default(0);
             $table->integer('votes')->default(0);
             $table->integer('answers')->unsigned()->default(0);
-            $table->integer('best_answer_id')->unsigned()->nullable();
+            $table->bigInteger('best_answer_id')->unsigned()->nullable();
             $table->timestamps();
         });
